requirements.txt
openai
langchain
langchain-community
faiss-cpu
streamlit
tiktoken
pdfplumber
python-dotenv
